:truck: Rename : Footer 컴포넌트 파일 및 폴더 명 변경
- Layout 폴더명을 Footer 로 수정하였습니다.
- Footer.tsx 파일명을 index.tsx 로 수정하였습니다.

Issues : #10
