+++ client/src/App.tsx
@@ -1,6 +1,6 @@
 import React from "react";
 import styled from "styled-components";
-import Footer from "./components/layouts/Footer";
+import Footer from "./components/Footer";
 
 function App() {
   return (
