+++ client/src/components/Footer/index.tsx
